+++ src/service/index.ts
@@ -2,3 +2,4 @@ export { default as goalService } from "./goalService";
 export { default as userService } from "./userService";
 export { default as cheeringMessageService } from "./cheeringMessageService";
 export { default as dailyAchievedHistoryService } from "./dailyAchievedHistoryService";
+export { default as mypageService } from "./mypageService";
